refactor: Moved API URL and movie fetching logic to separate files, refactored Proptypes
- Moved API URL to a separate `config.js` file to improve modularity of the code.
- Moved fetching movies logic from the API, which was previously in the `MainView` component, to a custom hook in a new `useMovies.js` file in the hooks directory.
- Simplified PropTypes in MovieDescription and MovieDirector component to only include relevant proptypes
