+++ src/config.js
@@ -1,0 +1,1 @@
+export const API_URL = 'https://cinephile-dc1b75a885d0.herokuapp.com';
